Extract keepAliveListener
